requirements.txt
mcp
Pillow
pyyaml
